refactor(game): move StormShipLogic to Weather
